LPS-62384 Move to NpmAnalyzerPlugin (No logical changes)
