+++ build.properties
@@ -42,8 +42,8 @@
             name=remoterepo;\
             phase=any;\
             type=R5,\
-        com.liferay.ant.bnd.bower.BowerAnalyzerPlugin,\
         com.liferay.ant.bnd.jsp.JspAnalyzerPlugin,\
+        com.liferay.ant.bnd.npm.NpmAnalyzerPlugin,\
         com.liferay.ant.bnd.sass.SassAnalyzerPlugin,\
         com.liferay.ant.bnd.spring.SpringDependencyAnalyzerPlugin
     -pluginpath=${project.dir}/tools/sdk/osgi/lib/plugin/bnd-repository.jar
